Add pageUtil and clean up PdfDocument
