+++ client/app/readerprototype/util/documentUtil.js
