+++ client/app/readerprototype/util/layoutUtil.js
