+++ client/app/readerprototype/util/readerConstants.js
